+++ acts_as_primary.gemspec
@@ -20,5 +20,6 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.7"
   spec.add_development_dependency "rake", "~> 10.0"
-  spec.add_development_dependency "rails", "~> 4.0"
+  spec.add_development_dependency "activesupport", "~> 4.0"
+  spec.add_development_dependency "activerecord", "~> 4.0"
 end
